Restructured the github tasks into separate files per task under a common subdirectory.  All class_paths remain the same with this structure :)
