+++ cumulusci/tasks/github/__init__.py
@@ -1,0 +1,5 @@
+from cumulusci.tasks.github.commit import CommitApexDocs
+from cumulusci.tasks.github.merge import MergeBranch
+from cumulusci.tasks.github.pull_request import PullRequests
+from cumulusci.tasks.github.release import CreateRelease
+from cumulusci.tasks.github.tag import CloneTag
